finished NoteViewModel, started NotePresenter & NoteState, created new package directory interface_adapters.note and refactored the files
