+++ src/interface_adapter/note/NotePresenter.java
@@ -1,0 +1,4 @@
+package interface_adapter.note;
+
+public class NotePresenter {
+}
